Refactor shared UI components into modules
